Run Bacon

Run the function rbacon:Baconrbacon::Bacon(...).

run_bacon(
  wdir,
  entity,
  postbomb = 0,
  cc = 1,
  seed = NA,
  alt_depths = NULL,
  quiet = FALSE,
  core = NULL,
  depths_eval = NULL,
  hiatuses = NULL,
  sample_ids = NULL,
  unknown_age = NULL,
  coredir = NULL,
  acc.mean = 20,
  acc.shape = 1.5,
  ssize = 2000,
  th0 = c(),
  thick = 5,
  p = NULL,
  ...
)

Arguments

wdir

Path where input files are stored.

entity

Name of the entity.

postbomb

Use a postbomb curve for negative (i.e. postbomb) 14C ages. 0 = none, 1 = NH1, 2 = NH2, 3 = NH3, 4 = SH1-2, 5 = SH3.

cc

Calibration curve for C-14 dates: cc = 1 for IntCal20 (northern hemisphere terrestrial), cc = 2 for Marine20 (marine), cc = 3 for SHCal20 (southern hemisphere terrestrial). For dates that are already on the cal BP scale use cc = 0.

seed

Set see to reproduce results. This seed is used for C++ executions, if it is not assigned then the seed is set by the system.

alt_depths

List of arrays with new depths.

quiet

Boolean to hide status messages.

core

Data frame with the core's data.

depths_eval

Numeric array with the sampling depths.

hiatuses

Data frame containing information of hiatuses.

sample_ids

Numeric array with IDs for the sampling depths.

unknown_age

Data frame containing information of unused ages.

coredir

Folder where the core's files core are and/or will be located.

acc.mean

The accumulation rate prior consists of a gamma distribution with two parameters. Its mean is set by acc.mean (default acc.mean=20 yr/cm (or whatever age or depth units are chosen), which can be changed to, e.g., 5, 10 or 50 for different kinds of deposits). Multiple values can be given in case of hiatuses or boundaries, e.g., Bacon(hiatus.depths=23, acc.mean=c(5,20)).

acc.shape

The prior for the accumulation rate consists of a gamma distribution with two parameters. Its shape is set by acc.shape (default acc.shape=1.5; higher values result in more peaked shapes).

ssize

The approximate amount of iterations to store at the end of the MCMC run. Default 2000; decrease for faster (but less reliable) runs or increase for cores where the MCMC mixing (panel at upper-left corner of age-model graph) appears problematic.

th0

Starting years for the MCMC iterations.

thick

Bacon will divide the core into sections of equal thickness specified by thick (default thick = 5).

p

progressor object from the package progressr.

...

Optional parameters for rbacon::Bacon.

Value

Saves MC ensemble, bacon_chronology and AM plot.

References

Blauuw, M. et al., Bayesian Analysis 6, 457-474 (2011)

Blauuw, M. et al., rbacon (2019), R package version 2.3.9.1

Comas-Bru, L. et al., SISALv2: A comprehensive speleothem isotope database with multiple age-depth models, Earth Syst. Sci. Data Discuss (2020) https://doi.org/10.5194/essd-2020-39, https://github.com/paleovar/SISAL.AM